zmg_backend/requirements.txt
django
djangorestframework
django-cors-headers
django-filter
pillow
djangorestframework-simplejwt
